Add Orbit intergration
Branding always needs Orbit, so it's a bit less effort to include
an OrbitClient in here rather than relying another library.

This also gives us the opportunity to update the OrbitClient, as the
one from rmp-orb-client is a little old and doesn't request Gzipped
responses or honor Orbit's cache lifetime preferences.